pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matomail"
version = "0.1.0"
description = "Local Gmail review assistant with LLM summaries and user-confirmed actions."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "google-api-python-client>=2.130",
    "google-auth-oauthlib>=1.2",
    "jinja2>=3.1",
    "openai>=1.30",
    "sqlalchemy>=2.0",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
]

[project.scripts]
matomail = "matomail.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
